added new controller and added annotations 19-08-2025

+++ DTO/ActiveClassicalCalendarDTO.cs
@@ -16,9 +16,9 @@ public class ActiveClassicalCalendarDTO
 
     public DateOnly BuyOrderExpiryDate { get; set; }
 
-    public double CallSellLTP { get; set; }
-    public double PutSellLTP { get; set; }
-    public double CallBuyLTP { get; set; }
-    public double PutBuyLTP { get; set; }
+    public decimal CallSellLTP { get; set; }
+    public decimal PutSellLTP { get; set; }
+    public decimal CallBuyLTP { get; set; }
+    public decimal PutBuyLTP { get; set; }
     public ClassicalCalendarStatus ClassicalCalendarStatus { get; set; }
 }
